app/requirements.txt
# Core dependencies for LightRAG (since we're using local source)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.11.0
networkx>=3.2
tiktoken>=0.5.0
aiofiles>=23.0.0
nest-asyncio>=1.5.0
pyuca>=1.2

# Gemini Integration
google-generativeai>=0.8.0

# Streamlit Frontend
streamlit>=1.28.0
streamlit-chat>=0.1.1
streamlit-echarts>=0.4.0

# Data Visualization & Analytics
plotly>=5.17.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Graph Visualization
pyvis>=0.3.2
python-louvain>=0.16

# Monitoring & Database
requests>=2.31.0

# Optional dependencies for enhanced storage (if needed)
# redis>=5.0.0
# neo4j>=5.13.0
# pymongo>=4.5.0
# qdrant-client>=1.6.0
# faiss-cpu>=1.7.4
# psycopg2-binary>=2.9.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
typing-extensions>=4.8.0
dataclasses-json>=0.6.0
html2text>=2020.1.16
lxml>=4.9.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0
python-multipart>=0.0.6 